refactor : 미기능 및 에러상태 수정
